refactor: upgrade device account in separate file
