+++ src/app/accounts/index.ts
@@ -21,6 +21,7 @@ export * from "./update-default-walletid"
 export * from "./update-display-currency"
 export * from "./username-available"
 export * from "./delete-business-map-info"
+export * from "./upgrade-device-account"
 
 const accounts = AccountsRepository()
